Moved AlterPersonality to the logicHelper file
